Plugin API for change and revision
Change [1] started to implement new API with RevisionApi.review.
This change implements the following API methods:

* abandon
* rebase
* restore
* revert
* submit
* delete (draft)

SSH review command was migrated to utilize the new API.
Acceptance tests were added for the new API methods.

[1] https://gerrit-review.googlesource.com/48664

+++ gerrit-server/src/test/java/com/google/gerrit/server/query/change/AbstractQueryChangesTest.java
@@ -606,12 +606,12 @@ public abstract class AbstractQueryChangesTest {
   }
 
   protected void assertResultEquals(Change expected, ChangeInfo actual) {
-    assertEquals(expected.getId().get(), actual._number);
+    assertEquals(expected.getId().get(), (int) actual._number);
   }
 
   protected void assertResultEquals(String message, Change expected,
       ChangeInfo actual) {
-    assertEquals(message, expected.getId().get(), actual._number);
+    assertEquals(message, expected.getId().get(), (int) actual._number);
   }
 
   protected TestRepository<InMemoryRepository> createProject(String name)
